backend/requirements.txt
asgiref
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-cors-headers==4.3.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
drf-yasg==1.21.7
filetype==1.2.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-isort==6.1.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isort==5.12.0
Markdown==3.4.4
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pdf2image==1.16.3
pep8-naming==0.13.3
Pillow==10.0.1
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==57.5.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
views-py==2.0.0
pycoin~=0.92.20230326
django-ckeditor==6.7.1
captcha~=0.5.0
django-simple-captcha~=0.6.0
bipwallet~=1.0.2